Implement sudo support
These changes allow to run pvpn as non-root user on the client system.
Running pvpn as root user requires a proper .ssh/config for the root
user, that does not require any user-home-relative paths to e.g. key
files.

Device creation is done separately, not by ssh. This allows to run ssh
without root privileges as long as the non-privileged user has
permissions to change tun device configurations. This is achived by the
user attribute of "ip tuntap":

    sudo ip tuntap add dev ${interface} mode tun user $(id -un)

For the server side a check has been implemented that checks wether the
connected user is either root or has sudo privileges.