Implement -b/--background and -k/--keep-devices CLI arguments
By default, pvpn will stay in the foreground. Whenever the script exits,
connections will be closed, statefiles will be deleted and the used tun
devices will be removed. Using -b/--background will start pvpn in the
former mode where ssh will run in the background and connetions have to
be closed with the -S/--stop argument.

By default, tun devices will be deleted as soon as a connection will be
stopped. Using the -k/--keep-devices will keep the tun device
configurations on the client and server.